sprint 11 (ver 1)

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,10 +3,8 @@ package ru.yandex.practicum.filmorate.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
-
 import java.util.Collection;
 
 
@@ -27,7 +25,7 @@ public class UserService {
             throw new NotFoundException("Неверный Id друга");
         }
         userStorage.addFriends(userId, friendId);
-    };
+    }
 
     public void deleteFriends(long userId, long friendId) {
         if (!userStorage.getUsers().containsKey(userId)) {
@@ -36,15 +34,18 @@ public class UserService {
         if (!userStorage.getUsers().containsKey(friendId)) {
             throw new NotFoundException("Неверный Id друга");
         }
-        userStorage.deleteFriends(userId, friendId);
-    };
+        if (userStorage.getUserFriendsIds().containsKey(userId) &&
+                userStorage.getUserFriendsIds().get(userId).contains(friendId)) {
+            userStorage.deleteFriends(userId, friendId);
+        }
+    }
 
     public Collection<User> findFriends(long userId) {
         if (!userStorage.getUsers().containsKey(userId)) {
             throw new NotFoundException("Неверный Id пользователя");
         }
         return userStorage.findFriends(userId);
-    };
+    }
 
     public Collection<User> findCommonFriends(long userId, long otherId) {
         if (!userStorage.getUsers().containsKey(userId)) {
@@ -54,8 +55,5 @@ public class UserService {
             throw new NotFoundException("Неверный Id друга");
         }
         return userStorage.findCommonFriends(userId, otherId);
-    };
-
-
-
+    }
 }
